requirements.txt
python>=3.9
torch==2.7.1
torchvision==0.22.1
torchaudio==2.7.1
tqdm>=4.67.1
numpy>=1.26.3
scipy>=1.13.1
pandas>=2.2.3
scikit-learn>=1.6.1
PyYAML>=6.0.2
rasterio>=1.4.3
scikit-image>=0.24.0
opencv-python>=4.11.0.86
tifffile>=2024.8.30
spectral>=0.24
albumentations>=2.0.5
einops>=0.8.1
matplotlib>=3.9.4
seaborn>=0.13.2
joblib>=1.4.2
colorama>=0.4.6